Enhance MeetingRoom and VideoTileCanvas components for improved layout and functionality; add useMeetingRoom hook for state management
